pyproject.toml
[project]
name = "awwbot"
version = "1.0.0"
description = "A Discord bot hosted on Vercel Functions, written in Python"
requires-python = ">=3.12"
dependencies = [
    "PyNaCl==1.5.0",
]
